# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateOrderItem:
    properties:
      product_name:
        example: Ноутбук
        type: string
      quantity:
        example: 1
        minimum: 1
        type: integer
    required:
    - product_name
    - quantity
    type: object
  dto.CreateOrderRequest:
    description: Запрос на создание нового заказа
    properties:
      delivery_address:
        example: ул. Пушкина, д. 10
        type: string
      items:
        items:
          $ref: '#/definitions/dto.CreateOrderItem'
        minItems: 1
        type: array
      user_id:
        example: 123
        type: integer
    required:
    - delivery_address
    - items
    - user_id
    type: object
  dto.CreateOrderResponse:
    description: Ответ после успешного создания заказа
    properties:
      message:
        example: Order created successfully
        type: string
      order:
        $ref: '#/definitions/entity.Order'
    type: object
...
  entity.GoodsItem:
    description: Товар в составе заказа
    properties:
      last_updated:
        example: 1694966400
        type: integer
      price:
        example: 15000
        type: number
      product_id:
        example: 789
        type: integer
      product_name:
        example: Ноутбук
        type: string
      quantity:
        example: 1
        type: integer
      total_price:
        example: 15000
        type: number
    type: object
  entity.Order:
    description: Информация о заказе
    properties:
      created_at:
        example: 1694966400
        type: integer
      delivery_address:
        example: ул. Пушкина, д. 10
        type: string
      driver_id:
        example: 456
        type: integer
      id:
        example: 1
        type: integer
      items:
        items:
          $ref: '#/definitions/entity.GoodsItem'
        type: array
      status:
        allOf:
        - $ref: '#/definitions/entity.OrderStatus'
        example: pending
      total_amount:
        example: 15000.5
        type: number
      user_id:
        example: 123
        type: integer
    type: object
  entity.OrderStatus:
    enum:
    - pending
    - confirmed
    - route_ready
    - assigned
    - in_progress
    - delivered
    - cancelled
    - failed
    type: string
    x-enum-comments:
      StatusAssigned: водитель назначен
      StatusCancelled: отменен
      StatusConfirmed: товар зарезервирован
      StatusDelivered: доставлен
      StatusFailed: ошибка
      StatusInProgress: в пути
      StatusPending: создан, ждет проверки склада
      StatusRouteReady: маршрут построен
    x-enum-descriptions:
    - создан, ждет проверки склада
    - товар зарезервирован
    - маршрут построен
    - водитель назначен
    - в пути
    - доставлен
    - отменен
    - ошибка
    x-enum-varnames:
    - StatusPending
    - StatusConfirmed
    - StatusRouteReady
    - StatusAssigned
    - StatusInProgress
    - StatusDelivered
    - StatusCancelled
    - StatusFailed
host: localhost:9091
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API для управления логистическими операциями включая аутентификацию,
    заказы, доставки и управление складом, построенный на базе нескольких микросервисов
    с использованием gRPC и REST.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Logistics Management API
  version: "1.0"
paths:
...
  /orders:
    get:
      description: Возвращает все заказы текущего авторизованного пользователя
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ
          schema:
            properties:
              orders:
                items:
                  $ref: '#/definitions/entity.Order'
                type: array
            type: object
        "500":
          description: Ошибка сервера
          schema:
            properties:
              error:
                type: string
              message:
                type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Получение списка заказов пользователя
      tags:
      - orders
    post:
      consumes:
      - application/json
      description: Создает новый заказ после проверки наличия товаров на складе
      parameters:
      - description: Данные для создания заказа
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CreateOrderResponse'
        "400":
          description: Некорректные данные или товара нет в наличии
          schema:
            properties:
              error:
                type: string
              message:
                type: string
            type: object
        "500":
          description: Ошибка сервера
          schema:
            properties:
              error:
                type: string
              message:
                type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Создание нового заказа
      tags:
      - orders
  /orders/{order_id}:
    get:
      description: Возвращает детальную информацию о конкретном заказе пользователя
      parameters:
      - description: ID заказа
        in: path
        name: order_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Детали заказа
          schema:
            $ref: '#/definitions/entity.Order'
        "400":
          description: Неверный ID заказа
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Ошибка сервера
          schema:
            properties:
              error:
                type: string
              message:
                type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Получение деталей заказа
      tags:
      - orders
  /orders/{order_id}/assign-driver:
    post:
      description: Назначает подходящего водителя на заказ и обновляет его статус
      parameters:
      - description: ID заказа
        in: path
        name: order_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешное назначение
          schema:
            properties:
              driver_id:
                format: int64
                type: integer
              message:
                type: string
              order_id:
                format: int64
                type: integer
              success:
                type: boolean
            type: object
        "400":
          description: Заказ не в pending статусе
          schema:
            properties:
              error:
                type: string
              message:
                type: string
            type: object
        "500":
          description: Ошибка сервера
          schema:
            properties:
              error:
                type: string
              message:
                type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Назначение водителя на заказ
      tags:
      - orders
...
securityDefinitions:
  BearerAuth:
    description: Введите 'Bearer ' followed by your JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"